#1016 Renamed utility calss 'SecurityTokenInfo' not to conflict with an entity of the same name.
